+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/ItemPocketRepository.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
  * @since 2025/6/8
  */
 @Repository
-public interface ItemPocketRepository extends ListPagingAndSortingRepository<ItemPocket, Integer> {
+public interface ItemPocketRepository extends ListPagingAndSortingRepository<ItemPocket, Long> {
 
   /**
    * 根据分类查找并返回对应的道具口袋数据
